feat(config): refactor configuration loading and remove unused properties
