+++ src/types/config.ts
@@ -20,8 +20,6 @@ export interface IConfig {
     };
     music: {
         enabled: boolean;
-        dashboard_url: string;
-        image: string;
         lavalink: {
             default_search: string;
             nodes: Array<{
